fix(DEV-11087): fix circular dependency
