+++ packages/sdk-react/src/components/approvalRequests/consts.ts
@@ -4,7 +4,12 @@ import HourglassEmptyIcon from '@mui/icons-material/HourglassEmpty';
 import PaidOutlinedIcon from '@mui/icons-material/PaidOutlined';
 import { ChipTypeMap, type SvgIcon } from '@mui/material';
 
-import { ApprovalRequestStatus } from './types';
+export enum ApprovalRequestStatus {
+  APPROVED = 'approved',
+  CANCELED = 'canceled',
+  REJECTED = 'rejected',
+  WAITING = 'waiting',
+}
 
 export const ROW_TO_STATUS_MUI_MAP: {
   [key in ApprovalRequestStatus]: ChipTypeMap['props']['color'];
