+++ packages/sdk-react/src/mocks/approvalRequests/approvalRequestsFixture.ts
@@ -22,6 +22,6 @@ export const approvalRequestsListFixture: Services['approvalRequests']['getAppro
       'approved',
       'rejected',
       'canceled',
-    ]) as components['schemas']['ApprovalRequestStatus'],
+    ]),
     user_ids: [entityUserByIdFixture.id, faker.string.nanoid()],
   }));
